Fix ElectronContainer getMainWindow() (#88)
Add support for providing option of main: true to window creation in Electron to specify the "main" window.  By default desktopJS will attach to the closed event for you and quit the application when the main window is closed.  To disable this specify quitOnClose: false in options.  If no window is specified as main, getMainWindow() will fallback to the first created window.

Fixes #87